finished consolidation
