--- cm.indexwin~/cmutil.c
@@ -1,50 +1,0 @@
-/*
- Circuit Music Labs utilities library file.
- Copyright (C) 2014  Matthias Müller - Circuit Music Labs
- 
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
- 
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
- 
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
- 
- circuit.music.labs@gmail.com
- 
- */
-
-#include "ext.h"
-#include <stdio.h>
-#include "cmutil.h"
-
-// RANDOM NUMBER GENERATOR (USE POINTERS FOR MORE EFFICIENCY)
-double cm_random(double *min, double *max) {
-		
-	return *min + ((*max - *min) * (((double)arc4random_uniform(RANDMAX)) / (double)RANDMAX));
-	
-}
-
-
-// LINEAR INTERPOLATION FUNCTION FOR BUFFER (passing float pointer)
-double cm_lininterp(double distance, float *buffer, t_atom_long b_channelcount, short channel) {
-	
-	long index = (long)distance; // get truncated index
-	distance -= (long)distance; // calculate fraction value for interpolation
-	
-	return buffer[index * b_channelcount + channel] + distance * (buffer[(index + 1) * b_channelcount + channel] - buffer[index * b_channelcount + channel]);
-}
-
-// LINEAR INTERPOLATION FUNCTION FOR WINDOW (passing douple pointer)
-double cm_lininterpwin(double distance, double *buffer, t_atom_long b_channelcount, short channel) {
-	
-	long index = (long)distance; // get truncated index
-	distance -= (long)distance; // calculate fraction value for interpolation
-	
-	return buffer[index * b_channelcount + channel] + distance * (buffer[(index + 1) * b_channelcount + channel] - buffer[index * b_channelcount + channel]);
-}
